Restore to app folder (#7)
* restored functionality in app folder

* aligned stitcher parameters with single class parameters, updated seam drawing, removed privacy of functions

* updated tutorial
