+++ apps/opencv_stitching_tool/opencv_stitching/.gitignore
@@ -1,0 +1,4 @@
+# python binary files
+*.pyc
+__pycache__
+.pylint*
