+++ apps/opencv_stitching_tool/opencv_stitching/stitching_error.py
@@ -1,0 +1,2 @@
+class StitchingError(Exception):
+    pass
