--- samples/python/stitching_detailed/.gitignore
@@ -1,4 +1,0 @@
-# python binary files
-*.pyc
-__pycache__
-.pylint*
