--- samples/python/stitching_detailed/stitching_error.py
@@ -1,2 +1,0 @@
-class StitchingError(Exception):
-    pass
