--- samples/python/stitching_detailed/test/.gitignore
@@ -1,13 +1,0 @@
-# Ignore everything
-*
-
-# But not these files...
-!.gitignore
-!test_matcher.py
-!test_stitcher.py
-!test_megapix_scaler.py
-!test_registration.py
-!test_composition.py
-!test_performance.py
-!stitching_detailed_old.py
-!SAMPLE_IMAGES_TO_DOWNLOAD.txt
