--- samples/python/stitching_detailed/tutorial/.gitignore
@@ -1,8 +1,0 @@
-# Ignore everything
-*
-
-# But not these files...
-!.gitignore
-!figures
-!Stitching Tutorial.ipynb
-!README.md
